+++ frontend/src/components/pages/admin/QuestionPage.tsx
@@ -3,11 +3,11 @@ import { DndProvider } from "react-dnd";
 import { HTML5Backend } from "react-dnd-html5-backend";
 import { Flex } from "@chakra-ui/react";
 
-import { HOME_PAGE } from "../../constants/Routes";
-import QuestionSidebar from "./QuestionSidebar";
-import QuestionEditor from "./QuestionEditor";
+import { HOME_PAGE } from "../../../constants/Routes";
+import QuestionSidebar from "../../question-creation/QuestionSidebar";
+import QuestionEditor from "../../question-creation/QuestionEditor";
 
-import { QuestionElement } from "../../types/QuestionTypes";
+import { QuestionElement } from "../../../types/QuestionTypes";
 
 const QuestionPage = (): React.ReactElement => {
   const [questionElements, setQuestionElements] = React.useState<
